Pivotal ID # 188131969: Archive submission request (#873)
https://www.pivotaltracker.com/story/show/188131969
- updated to use document

+++ submission/persistence-mongo/src/main/kotlin/ac/uk/ebi/biostd/persistence/doc/db/reactive/repositories/Repositories.kt
@@ -157,7 +157,7 @@ interface SubmissionRequestFilesRepository : CoroutineCrudRepository<DocSubmissi
     suspend fun countByAccNoAndVersion(
         accNo: String,
         version: Int,
-    ): Int
+    ): Long
 
     @Query("{ 'accNo': ?0, 'version': ?1, 'status': ?2 }")
     @Meta(flags = [CursorOption.NO_TIMEOUT])
